Optimized Commerce's fetchPlaceholders (#380)
* reduces cognitive complexity

* refactor: enhance placeholder fetching logic to consolidate data handling and improve clarity

* refactor: migrate placeholder fetching logic from aem.js to commerce.js for improved modularity
